hasPlayedOnce_ flag
If isPlaying_ is true but hasPlayedOnce_ is false, it implies that this is the first time this video is playing

If this is the first play cycle, we disregard any restart requests.

Once a video has played, we set hasPlayedOnce_ to true so that in any subsequent play cycles, we do honor the restart request.